requirements.txt
appdirs==1.4.4
asgiref==3.3.4
distlib==0.3.1
Django==3.2
django-crispy-forms==1.11.2
django-filter==2.4.0
filelock==3.0.12
Pillow==8.2.0
pytz==2021.1
six==1.15.0
sqlparse==0.4.1
virtualenv==20.4.3
